requirement.txt
aiohttp==3.7.4.post0
aiosqlite==0.17.0
asgiref==3.5.0
async-timeout==3.0.1
attrs==21.4.0
chardet==4.0.0
discord-ext-menus==1.0.0a32+g8686b5d
discord.py==1.7.3
idna==3.3
multidict==6.0.2
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
six==1.16.0
sqlparse==0.4.2
typing_extensions==4.2.0
tzdata==2022.1
tzlocal==4.2
yarl==1.7.2
